# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Eltuvchi API",
        "description": "Eltuvchi platformasi uchun API hujjatlari",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local server"
        }
    ],
    "paths": {
...
        "/api/super-admin/drivers": {
            "get": {
                "tags": [
                    "Admin Drivers"
                ],
                "summary": "Get paginated list of drivers",
                "operationId": "1a7efeb46368c80dd58fda2377a45c5e",
                "responses": {
                    "200": {
                        "description": "Paginated list of drivers",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/AdminDriver"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin Drivers"
                ],
                "summary": "Create a new driver",
                "operationId": "05e1037936256c29e5474f7ac783d22f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "type": "integer"
                                    },
                                    "status": {
                                        "type": "string"
                                    },
                                    "balance": {
                                        "type": "integer"
                                    },
                                    "points": {
                                        "type": "integer"
                                    },
                                    "details": {
                                        "type": "object"
                                    },
                                    "settings": {
                                        "type": "object"
                                    },
                                    "taxopark_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Driver created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AdminDriver"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/super-admin/drivers/{driver}": {
            "get": {
                "tags": [
                    "Admin Drivers"
                ],
                "summary": "Get a single driver",
                "operationId": "2a40015edc545ae9a8f51bb42def1fd5",
                "parameters": [
                    {
                        "name": "driver",
                        "in": "path",
                        "description": "Driver ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Driver details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AdminDriver"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Driver not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Admin Drivers"
                ],
                "summary": "Update a driver",
                "operationId": "cf3e76113df16e600dcd6a421be80880",
                "parameters": [
                    {
                        "name": "driver",
                        "in": "path",
                        "description": "Driver ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "status": {
                                        "type": "string"
                                    },
                                    "balance": {
                                        "type": "integer"
                                    },
                                    "points": {
                                        "type": "integer"
                                    },
                                    "details": {
                                        "type": "object"
                                    },
                                    "settings": {
                                        "type": "object"
                                    },
                                    "taxopark_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Driver updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AdminDriver"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Driver not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin Drivers"
                ],
                "summary": "Mark driver as inactive",
                "operationId": "879eed17c16571def4ed93df347045ff",
                "parameters": [
                    {
                        "name": "driver",
                        "in": "path",
                        "description": "Driver ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Driver marked as inactive",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AdminDriver"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Driver not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "AdminDriver": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "status": {
                        "type": "string"
                    },
                    "balance": {
                        "type": "integer"
                    },
                    "points": {
                        "type": "integer"
                    },
                    "details": {
                        "type": "object"
                    },
                    "settings": {
                        "type": "object"
                    },
                    "user": {
                        "type": "object"
                    },
                    "taxopark": {
                        "type": "object"
                    },
                    "orders_count": {
                        "type": "integer"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Sanctum authentication using Bearer token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}